@@ -1,8 +1,7 @@
 import * as authActions from '../auth/_store/actions';
 import * as categoriesActions from '../categories/_store/actions';
-import * as modalActions from '../modal/_store/actions';
 import * as payconiqActions from '../payconiq/_store/actions';
 import * as orderActions from '../order/_store/actions';
 import * as ordersActions from '../orders/_store/actions';
 
-export { authActions, categoriesActions, modalActions, orderActions, ordersActions, payconiqActions };
+export { authActions, categoriesActions, orderActions, ordersActions, payconiqActions };
